requirements.txt
fitlog==0.5.8
tqdm==4.26.0
seaborn==0.9.0
opencv_python==4.1.0.25
easydict==1.9
gluoncv==0.7.0
numpy==1.18.5
scipy==1.1.0
mxnet_cu101==1.5.0
matplotlib==2.2.3
Pillow==7.1.2
typing==3.7.4.1
